requirements-gpu.txt
absl-py==2.3.1
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
astunparse==1.6.3
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
Cython==3.1.3
darkdetect==0.8.0
easydict==1.13
ffmpeg-python==0.2.0
filelock==3.19.1
filterpy==1.4.5
fire==0.7.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.2
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mysql-connector-python==9.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
onnx==1.19.0
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.5.5.64
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pefile==2024.8.26
pillow==11.3.0
playsound==1.2.2
polars==1.32.3
prettytable==3.16.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pygrabber==0.2
PyMySQL==1.1.2
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
simsimd==6.5.1
six==1.17.0
soupsieve==2.7
stringzilla==3.12.6
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.3
xlwings==0.33.15
